Render components as component files
