+++ src/components/Fields/RichText/Element.js
@@ -1,5 +1,4 @@
 /* eslint-disable react/jsx-props-no-spreading */
-/* eslint-disable no-param-reassign */
 import React from 'react';
 import PropTypes from 'prop-types';
 
@@ -39,42 +38,4 @@ Element.defaultProps = {
   children: null,
 };
 
-const withMarks = (content, leaf) => {
-  if (leaf.bold) {
-    if (leaf.italic) {
-      return <strong><em>{content}</em></strong>;
-    }
-
-    return <strong>{content}</strong>;
-  }
-
-  if (leaf.italic) {
-    return <em>{content}</em>;
-  }
-
-  return content;
-};
-
-const Leaf = ({ attributes, children, leaf }) => (
-  <span {...attributes}>{withMarks(children, leaf)}</span>
-);
-
-Leaf.propTypes = {
-  attributes: PropTypes.object,
-  children: PropTypes.node,
-  leaf: PropTypes.shape({
-    bold: PropTypes.bool,
-    italic: PropTypes.bool,
-    underline: PropTypes.bool,
-  }).isRequired,
-};
-
-Leaf.defaultProps = {
-  attributes: {},
-  children: null,
-};
-
-export {
-  Element,
-  Leaf,
-};
+export default Element;
